pyproject.toml
[tool.poetry]
name = "med_clinics"
version = "0.1.0"
description = ""
authors = ["Javokhir <kasimovjavohir01@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Pillow = "^9.1.0"
psycopg2 = "^2.9.3"
django-cors-headers = "^3.11.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
